+++ src/components/elements/link/link.ts
@@ -1,5 +1,5 @@
 import './link.css';
-import Component from '../component';
+import Component from '../../component';
 
 interface LinkOptions {
   className: string;
